+++ src/modules/customer/get-balance/get-balance.mock.ts
@@ -1,0 +1,21 @@
+export const request = {
+  "headers": {
+    "Content-Type": "application/json",
+    "Authorization": "Bearer bearer-token"
+  },
+};
+export const response = {
+    "body": {
+      "month_to_date_balance": "23.44",
+      "account_balance": "12.23",
+      "month_to_date_usage": "11.21",
+      "generated_at": "2019-07-09T15:01:12Z"
+    },
+    "headers": {
+      "content-type": "application/json; charset=utf-8",
+      "status": 200,
+      "ratelimit-limit": 1200,
+      "ratelimit-remaining": 1137,
+      "ratelimit-reset": 1415984218
+    },
+};