+++ src/modules/customer/index.ts
@@ -1,0 +1,2 @@
+export * from './get-balance/get-balance';
+export * from './types';