+++ src/modules/customer/types/balance.ts
@@ -1,0 +1,6 @@
+export interface IBalance {
+  month_to_date_balance: string;
+  account_balance: string;
+  month_to_date_usage: string;
+  generated_at: string;
+}